TINY-9184: Combinations
